Fix possible race condition around file locks
`FileLock#lock` doesn't always block until it can acquire a lock. If the lock
is held by the same JVM, it just throws an `OverlappingFileLockException`.
To work around that, this replaces a call to `FileLock#lock` by a while loop
that calls `FileLock#tryLock`, that stops when the lock is actually acquired.