Android/hilt unit tests (#5169)
## Summary

Prove that we can run the architecture-sample unit tests

The major feature missing here is to allow for kotlin modules to mark
any internal modules they depend on as friend modules (there was already
an isolated case for KotlinTests and the main source code) . Following
how the shared-tests of architecture samples depends on app main module
and in turn tests depend on shared-tests and main sources (as they do) ,
I've covered the missing gaps to get everything to work properly.

## Features added

- Kotlin `-X-friend-paths` is now used depending on transitive module
deps, for AndroidKotlinModules (this can later be expanded to
KotlinModule too)
- An androidNamespace is required for internal libs to function
- Android manifest for internal libs/modules needs to add metadata to
the manifest too
- AndroidHiltSupport should also be available to all android modules,
not just apps

## Fixes

- The AndroidHiltTransformAsm was dropping the META-INF, but it is
required for `-X-friend-paths` to function , as the kotlin module
contains the symbols necessary for internal members to be available to
the depending modules
- Ironed out the usage of androidApplicationNamespace vs
androidApplicationId as their usage becomes clearer to me the more apps
we integrate - this is under development and I expect their usage to be
stabilised once build variants are added too

## Unlocked workflows

This PR proves the following to be working:
- Complex unit test setup with Hilt
- A shared android module (shared-tests) can be given as a dependency
(it's a dependency of `app.test` and `app.androidTest` which will be
wired in in a follow up PR)
- Android Kotlin internal and top level members can be used between
depended modules (app.test can use app top level members as app compiled
path is passed as a friend path)

## Previous work reference

This is a continuous development for supporting android worfklows using
Hilt and by extension Kotlin KSP features.

Previous related work items:
- https://github.com/com-lihaoyi/mill/pull/4759
- https://github.com/com-lihaoyi/mill/pull/4485
- https://github.com/com-lihaoyi/mill/pull/4557

## Kotlin build issue

Need to check why arrow is not building, but probably I'll move the
friend paths functionality to AndroidKotlinModule if I don't manage to
identify the issue for the time being

Due to the use of outer in KotlinTest
```
    override def kotlincOptions: T[Seq[String]] = Task {
      outer.kotlincOptions().filterNot(_.startsWith("-Xcommon-sources")) ++
        Seq(s"-Xfriend-paths=${outer.compile().classes.path.toString()}")
    }
```

The `-Xfriend-paths` are picked from the parent and not computed
properly (so the module that a kotlin test directly depends on is
missing. e.g.

```
-Xfriend-paths=/home/vnicolaou/mill/out/example/thirdparty/arrow/downloadedRepo.dest/arrow-bc9bf92cc98e01c21bdd2bf8640cf7db0f97204a/out/arrow-libs/core/arrow-annotations/jvm/compile.dest/classes,/home/vnicolaou/mill/out/example/thirdparty/arrow/downloadedRepo.dest/arrow-bc9bf92cc98e01c21bdd2bf8640cf7db0f97204a/out/arrow-libs/core/arrow-atomic/jvm/compile.dest/classes,/home/vnicolaou/mill/out/example/thirdparty/arrow/downloadedRepo.dest/arrow-bc9bf92cc98e01c21bdd2bf8640cf7db0f97204a/out/arrow-libs/core/arrow-core/jvm/compile.dest/classes
```

but the kotlinc friend paths are from the outer module

```
[
  "-Xfriend-paths=/home/vnicolaou/mill/out/example/thirdparty/arrow/downloadedRepo.dest/arrow-bc9bf92cc98e01c21bdd2bf8640cf7db0f97204a/out/arrow-libs/core/arrow-annotations/jvm/compile.dest/classes,/home/vnicolaou/mill/out/example/thirdparty/arrow/downloadedRepo.dest/arrow-bc9bf92cc98e01c21bdd2bf8640cf7db0f97204a/out/arrow-libs/core/arrow-atomic/jvm/compile.dest/classes,/home/vnicolaou/mill/out/example/thirdparty/arrow/downloadedRepo.dest/arrow-bc9bf92cc98e01c21bdd2bf8640cf7db0f97204a/out/arrow-libs/core/arrow-core/jvm/compile.dest/classes",
  "-Xexpect-actual-classes"
]
```

I'm not gonna handle this in this PR, instead I've moved this logic to
AndroidKotlinModule and we can deal with it at a later stage

Pull request: https://github.com/com-lihaoyi/mill/pull/5169