Factor thread pool used to run tasks (#5198)
This makes `mill.exec.Execution` accept a thread pool in its
constructor, rather than creating one every time tasks need to be run.
In BSP and in watch mode, this should save the thread pool creation
across task runs. I don't have numbers to back the impact of this
change, but factoring the thread pool was part of [a bunch of
optimizations](https://github.com/coursier/sbt-coursier/pull/256) in the
coursier integration in sbt a few years ago. Plus given we follow more
of a single-execution-at-a-time model, there's no reason not to re-use
the thread pool across executions.