Fix: tests loading resources don't work in Intellij triggered tests with BSP import (#5202)
Tests that load resources do not work in some BSP clients like Intellij
IDEA (https://github.com/com-lihaoyi/mill/issues/4427) because these BSP
clients do not include resources classpath when running the tests from
the IDE.

Both sbt and maven (and presumably gradle) copy the resources into the
compile destination directory, so while it seems like a hack (as the
resources can in theory clash with the compiled classes), this seems to
be a working solution.

We work around this by detecting if the BSP client is Intellij and
performing the merging of the `compile` and `resources` outputs into a
separate folder (`bspBuildTargetCompileMerged`) for it.

Additionally, if `bspBuildTargetCompileMerged` is invoked, it invokes
the `bspBuildTargetCompileMerged` of all the transitive dependencies and
uses the `dest` dir of `bspBuildTargetCompileMerged` as a part of test
runner classpath.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>