skip broken scalalib migrating tests on windows (#5218)
Going to just disable the scalalib sbt migration example tests on
windows for now. They pass when run alone on CI
(https://github.com/com-lihaoyi/mill/actions/runs/15302931852/job/43048053802),
so this is probably just a test infra problem. Although disabling them
loses us a bit of coverage, this isn't actually a user-facing problem,
so we'll just go with that for now