Fix `ClassNotFoundException` when testing Scala.js module on Scala 2.12  (#5226)
Hit this porting the com.lihaoyi libraries to Mill 1.0.0-RC1. This was
broken in https://github.com/com-lihaoyi/mill/pull/4784, which causes
the module's resolver's `mapDependencies` to resolve scala-library
2.12.x on modules with that scala version, which conflicts with the rest
of the worker and tools classpath which is on Mill's scala version
(previously 2.13, now 3.x). This PR uses the `jvmWorker`'s resolver
instead that does not force any Scala version and thus allows a
consistent set of dependencies to be resolved

Not sure why the bug doesn't break the unit tests, probably due to a
different classloading environment than the distribution. Tested
manually with `dist.installLocal` and it fixes the hang in the
`sourcecode` and `mainargs` test suites that I was seeing with 1.0.0-RC1