Add explicit dependency on gson not to depend on an interval
ch.epfl.scala:bsp4j:2.2.0-M2 depends on org.eclipse.lsp4j:org.eclipse.lsp4j.jsonrpc:0.20.1,
that itself depends on com.google.code.gson:gson:[2.9.1,2.11)

In order not to depend on a version interval and make the resolution reproducible,
we add an explicit dependency on a gson version in this interval.

Depending on version intervals makes resolution check for new versions by pulling
version listing repeatedly, even when all artifacts are in cache. Depending on
a fixed version changes that, and makes resolution not attempt to download anything
if all artifacts are in cache.