Fix intellij scalac settings (#5220)
This fixes an issue that the translation from mill module names to
intelliJ module names sometimes uses `.` as a delimiter and sometimes
`-`. Which causes intelliJ not to be able to link certain settings to
the module.

The issue was that at one point all `Segment.Cross` were translated into
`Segment.Label`. And those two use a different delimiter.
```
  def moduleName(p: Segments): String =
    p.value
      .foldLeft(new StringBuilder()) {
        case (sb, Segment.Label(s)) if sb.isEmpty => sb.append(s)
        case (sb, Segment.Cross(s)) if sb.isEmpty => sb.append(s.mkString("-"))
        case (sb, Segment.Label(s)) => sb.append(".").append(s)
        case (sb, Segment.Cross(s)) => sb.append("-").append(s.mkString("-"))
      }
      .mkString
      .toLowerCase()
```

This fix might be quite intrusive for existing projects as it renames
the actual modules. The intended delimiter `-` was not used in most
places. This could be avoided by using the delimiter `.` for cross
modules as well.

I tested it with
[this](https://github.com/com-lihaoyi/mill/commit/c1f5d6d5717899c0cc3d139c1b07c2ef846c08a1)
setup. (Are there any automated tests for genIdea?)

Pull request: https://github.com/com-lihaoyi/mill/pull/5220