Exit client early if the daemon fails to start (#5267)
This makes the Mill client print the daemon stdout and stderr if the
daemon crashes before we even connect to it (and redirect its streams).
This has been useful to me when hacking on
https://github.com/com-lihaoyi/mill/pull/5155, but this should also be
useful if anything goes silently wrong when fetching the daemon class
path, and it fails to start early because of that.

Currently, in such a scenario, the client keep waiting indefinitely, and
users have no idea what's going on. With this PR, the client prints
things like this straightaway, which should help users debug things or
report an issue:
```text
Mill daemon exited unexpectedly!
No daemon stdout

Daemon stderr:

Exception in thread "main" java.lang.RuntimeException: Mill daemon early crash requested
	at scala.sys.package$.error(package.scala:27)
	at mill.daemon.MillDaemonMain$.main(MillDaemonMain.scala:15)
	at mill.daemon.MillDaemonMain.main(MillDaemonMain.scala)
```