Move internal task into dedicated `GenIDeaModuleInternal` module
This uses the same technique of an implicitly attached sub-module to avoid adding private and likely changing internals to public stable modules. this follows the previous BSP refactorings.

This is just the refactoring. There are likely more improvement that can be applied, now that we are able to cache in the context of a module.

Removed the `GenIdeaModule.intellijModulePath` in favor to the redundant `intellijModulePathJava`. This should reduce confusion  about what to override. But it needs a deprecation in Mill 0.12 which I'll add in a separate PR.