Android/rework resources (#5285)
## Motivation

This is driven from trying to integrate
[compose-samples](https://github.com/android/compose-samples) in mill.
There are a few issues which boil down to managing android resources and
handling run classpath (along with run dependencies).

This PR is a small part of reworking how we manage android resources and
creates a baseline to work on (i.e.. the separation + working tests)

## Removal from generatedSources

Having androidRes in generatedSources always seemed to be problematic
and we had to do some code gymnastics for the proper classpath
resolution and packaging (see other cleanup changes for an idea).
Instead of having it as generated sources, I've created a new
compilation path, that's added to the compile classpath.

Another reason, is that we'll need to separate resources from compile
classpath and run classpath (as I've discovered by playing around with
compose samples), so having them be compiled with the rest of the
sources and added into the global compile destination creates management
problems as we don't want to package all of them.

This could also be a benefit of somewhat incremental compilation, as
android resources are compiled independently of the all sources, and are
passed to the compile stage as part of the classpath. But this wasn't
the main reason to do it.

This also is similar to how AGP works


![image](https://github.com/user-attachments/assets/735f567e-4863-4bb2-a0e6-197943a286c6)


![image](https://github.com/user-attachments/assets/5e139d0e-4d09-4b57-8fec-e8a2dab01375)


## Other cleanups

- `android:debuggable` is not needed anymore apart from the lint example
for the lint purposes
- AndroidBuildConfig no longer needs to call the androidRClasses
explicitly as they have been removed from the generatedSources (this was
a temporary hack on my part for the purposes of getting Hilt to work),
so it now instead works with Hilt and KSP generated sources out of the
box with a call to super
- AndroidBuildConfig had duplicated references with AndroidAppModule,
now removed
- Empty generatedSources workarounds in instrumented tests and android
app tests are no longer necessary

## Future

So, it looks like a lot is happening in AGP for androidResource. For
one, AGP has its own [merging algorithm
](https://android.googlesource.com/platform/tools/base/+/refs/heads/mirror-goog-studio-main/sdk-common/src/main/java/com/android/ide/common/resources/)
, which at first glance, we could reuse, if we find an easy entry point
(it's also present in the android sdk cmdline-tools). Gradle usage is
[here](https://android.googlesource.com/platform/tools/base/+/refs/heads/mirror-goog-studio-main/build-system/gradle-core/src/main/java/com/android/build/gradle/tasks/MergeResources.kt)


![image](https://github.com/user-attachments/assets/38a5db33-4f07-458b-86a9-a957e9a9a96a)


Then, there are a few issues for missing R values which we haven't hit
with the todo example, but we hit with compose samples. These R values
come from the run classpath (we currently use compile classpath), and to
use that we need multiple changes:
1. Resolved Run Mvn Deps with proper android configuration to get all
the run dependencies
2. Compute a proper run classpath (I've already done some attempts with
mixed success)
3. Package the app's res against the run classpath res instead of the
compile res

My plan is to deliver these steps gradually and always testing against
compose samples to see if I'm going in the right direction. My
compose-samples branch is
[here](https://github.com/vaslabs-ltd/mill/tree/android/compose-samples-jetLagged)
, which is basically a playground instead of something I'm planning to
deliver, but can give an idea of where things are going.