Add tests for semanticdb generation in BSP server and make them work for build.mill (#5311)
This adds non-regression tests for semanticdb generation in the BSP
server, by enabling the generation of semanticdb in the BSP integration
tests. This also make sure these are generated for generated sources in
the build (like `build.mill`), by creating the semanticdb of original
sources (like `build.mill`) from the semanticdb of generated ones, like
Scala CLI does.

This also tweaks related things: this ensure scala-specific BSP requests
only return results for Scala modules, and restricts job parallelism in
the BSP integration tests to one job at a time, so that the BSP logs,
which are validated, are more reproducible. This is necessary when
enabling semanticdb generation, as compilation happens twice in that
case (once for classes, and a second time for semanticdbs), and their
output might be mixed without this.