Fix missing last char of log lines that exactly match terminal width (#5336)
Fixes https://github.com/com-lihaoyi/mill/issues/5334

The basic issue is that "lines that exactly fill the terminal" are
special-cased: the visual cursor is put on the previous line, but the
logical cursor remains on the end of the current line, and so "clear to
end of line" ends up wiping out the last character (e.g. see
https://github.com/mattiase/wraptest). The solution for this is to
"clear to end of line" _before_ writing it, so that we never clear at
the of the line

Tested this on my windows laptop, no noticeable flickering, and the
last-char rendering seems to be fixed. Presumably that's because even
though we are clearing the individual lines before writing them, we
never clear the entire prompt, and so the prompt doesn't flicker like it
did in the past