Add `--no-filesystem-checker` flag to globally disable filesystem read/write checks (#5278)
Fixes https://github.com/com-lihaoyi/mill/issues/5248

I kept all the `os.Checker`s in place and made them check a threadlocal
to see whether to perform a check or not, rather than swapping out the
`Checker` based on the flag. This is necessary to support cases like
`BuildFileCls` where plumbing in the flag is not easy.

Updated the `os-checker` integration tests to cover this, and mentioned
it in the sandbox example tests