Clean up Mill CLI flags (#5343)
I went and marked the various "internal" flags as `hidden = true`:
things like `--bsp`, `--tab-complete`, and so on aren't really intended
to be used directly by users, so no need to show them in the CLI
`--help` message. Others like `--no-build-lock` or
`--no-wait-for-build-lock` are kind of advanced features

All the internal and advanced flags no longer show in the normal
`--help` output, but instead in the output of `--help-advanced` which is
mentioned in the `--help` message so people can find it.

Deprecated flags are no longer shown in either `--help` or
`--help-advanced`, as the assumption is that only legacy scripted
callsites will hit these and anyone interactive should have no reason to
use them:

- `--no-server` is replaced by `--no-daemon`
- `--enable-ticker` and `--disable-ticker` are replaced by `--ticker
true/false`
- `--home` is unused
- `--repl` is not supported
- `--disable-callgraph` no longer works
- `--disable-prompt` no longer works
- `--silent` seems to have been unused for quite some time

Fixes https://github.com/com-lihaoyi/mill/issues/4938 and
https://github.com/com-lihaoyi/mill/issues/2158 and
https://github.com/com-lihaoyi/mill/issues/1946