Support command overloading in a binary compatible fashion (#5376)
Fixes  https://github.com/com-lihaoyi/mill/issues/5281

The previous code already collected all command overloads in the
`Discover` data structure, but it returned one arbitrarily when queried.
This PR makes it return the one with the longer argument list, which
supports `@unroll`-style signature evolution of adding parameters with
default values to the right side of a parameter list.

Added a unit test that fails on main, passes on this PR

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>