Fix VcsVersion short hash collision in Mill website publishing (#5388)
Fix the problem by picking the latest commit matching the VcsVersion
short hash, to avoid an ambiguous hash error on checkout.

Short commit SHAs always have a risk of collision, even those generate
by `git rev-parse --short`, since any short SHA that is currently unique
may collide with a commit that is added in future. So the best we can do
is to pick the most recent SHA which is most likely the right one (with
a 1/16million chance of collision for a length-6 short SHA)

Tested locally by using ` ./mill -i website.githubPages`

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>