Combine `mill.define` and `mill.api` into same `mill.api` namespace (#5407)
- `core/api/` has been moved to `core/api/shared/`, under
`mill.api.shared`
- `core/define/`  has been moved to `core/api/`, under `mill.api`
- `mill.api` now has explicit exports for the relevant user-facing
classes in `mill.api.shared`

The goal of this is to eliminate the user-facing split between
`mill.define` and `mill.api`. Previously, the split was purely based on
the internal concerns of what classes are necessary to interface with
`mill.daemon`, and is totally irrelevant to the user. This causes
confusion in where to find things, since it's generally impossible to
remember what lives in which one

This change unifies the user-facing API under `mill.api`, while keeping
the backend split of `core/api` and `core/api/shared` to accommodate the
separation necessary internally. Now users can directly import stuff
from `mill.api` all the time, and whether the actual implementation
lives in `mill.api` or `mill.api.shared` doesn't really matter since
everything that lives elsewhere is re-exported

For now the number of things exported from `import mill.api._` and
`import mill._` is somewhat arbitrary, but we can add more exports later
if necessary