Android: added `tools` sdk package to install automatically (#5437)
### Motivation
Running the android tests on a vanilla setup lead to the error of not
having `proguard` and provided `proguard` default configurations which
the `r8` third party examples rely on.

### Solution
Added `tools` to the sdk packages that get installed automatically and
removed it from the setup android action

### Thoughts for the Future
I think the future goal could be for mill to setup everything (with
automatic license accept on CI) and we can slowly remove the setup
android steps on the CI.
For example `cmdline-tools`'s manual installation seems necessary for
now since sdkmanager comes with it.