Park cursor after the multiline prompt rather than at the start (#5452)
This would allow the prompt to continue to be shown after using `Ctrl-C`
to interrupt a command, and avoid messing up the prompt in terminals
like Windows Terminal which do not clear the screen after `Ctrl-C`

Instead of bundling the `up` codes immediately after every prompt and of
the same height, we instead print the `up` codes before each prompt
equal to the height of the _previously rendered_ prompt, which we track
in a `lastPromptHeight` variable. This leaves the cursor after the
prompt in between prints, so on interruption the prompt remains visible
above.

Updated the unit tests to account for the cursor being on a new blank
line, and manually tested it using `rm -rf out && ./mill-assembly.jar -i
__.compile` on the Mill repo