Fix `gpg` hanging waiting for input when more than one key is present in keychain (#5471)
Mill's artifact signing for publishing needs `gpg` to import the signing
keys. However, the current implementation naively assumes that the key
it imports is the only key.

If you have other keys in your keychain, a couple of things could
happen:

- `gpg` by default signs using the first key in your keychain, which
isn't necessarily the key you just imported, so your artifact can be
signed using the wrong key.
- If the key `gpg` picks has a passphrase, `gpg` will hang indefinitely
waiting for you to enter the passphrase to stdin, which will never
happen, as it's mill handling the stdin.

This PR ensures that when the key is imported into GPG, its ID is
extracted and then used in the signing commands to specify which key to
use.

Tested manually on Linux and Mac.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>