Get coursier default repositories only through Resolve.defaultRepositories
This now respects mirrors and all, so that we don't need to use
`Resolve().finalRepositories` for that.

In later commits, at some places, we get those default repositories
via CoursierConfigModule tasks instead, which ensures we respect any change
in the env vars and Java properties that are used to compute those
repositories.