Reapply "Fix `gpg` hanging waiting for input when more than one key is present in keychain" (#5503)
This reapplies https://github.com/com-lihaoyi/mill/pull/5471.

There was a [problem with publishing
builds](https://github.com/com-lihaoyi/mill/actions/runs/16190480746/job/45704884047)
when this PR and sonatype snapshot publishing interacted together.

The issue stemmed from the fact that `"".split(',')` produces
`Array("")`, which then were interpreted as GPG arguments.

This PR adds additional logging of gpg commands, making sure that
passphrases are not leaked to logs. If user provides their own GPG
arguments, only the argument count and the source is printed, as we do
not know which of the provided arguments are sensitive.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>