Fix and add tests to sonatype snapshot publishing (#5525)
Unfortunately, due to moving a bunch of files around, Sonatype Central
snapshot publishing is broken in mill 1.0.0.

We didn't catch that because CI didn't run any tests.

This PR fixes the breakage (which was caused by using the wrong string
to refer to Maven worker's class) and adds the tests that publish to a
local directory.

Fixes https://github.com/com-lihaoyi/mill/issues/5513

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>