Android compose samples: JetNews (#5495)
Added
[JetNews](https://github.com/android/compose-samples/tree/main/JetNews)
to the compose samples.
<img width="1920" height="1080" alt="image"
src="https://github.com/user-attachments/assets/dab31d7f-6604-44c9-bf52-0e498ee1699c"
/>

The object structure is similar to the JetLagged sample. R8 still
currently throws some errors but converting them to warnings allows the
app to function and the tests to pass successfully.
This expands the test coverage and will be useful for detecting
regressions in the future

---------

Co-authored-by: Vasilis Nicolaou <vasilis.nicolaou@vaslabs.io>