Re-use Mill daemon in example/integration tests run in `.local` mode (#5548)
Cuts the time taken for `./mill -ik 'example.javalib.__.local.daemon'`
on my laptop from ~260s to ~150s, with similar speedups for many of the
other applicable test suites.

This change reduces the hermeticity of the tests to which it is applied,
in exchange for performance. Some code changes were necessary in order
to make the Mill daemon stable enough for this to work, but presumably
those were pre-existing bugs that should be fixed anyway.

A lot of the `integration` tests and `example.migrating` tests are not
amenable to re-using the Mill daemon for a variety of reasons, so those
are left as `.packaged` and `.native`. These typically exercise the edge
cases in Mill's daemon and lifecycle management, whereas the example
tests on `.local` mode tend to be exercising features that are unrelated
to the method of executing Mill