Snapshot handling in `SonatypeCentralPublishModule/publishAll` (#5539)
Currently, there are two ways to publish artifacts to Sonatype Central:
- `./mill __.publishSonatypeCentral`
- `./mill SonatypeCentralPublishModule/publishAll`

They have different code paths, and when adding snapshot support, I only
added it to `publishSonatypeCentral`, which resulted in people being
confused about why things do not work when they use
`SonatypeCentralPublishModule/publishAll`.

I looked at whether snapshot support can be added to `publishAll`, but
it seems it is a bit of a hack, as the `publishArtifacts:
mill.util.Tasks[PublishModule.PublishData] =
Tasks.resolveMainDefault("__:PublishModule.publishArtifacts")` value
does not have enough metadata for snapshot publishing (we require
`PublishInfo`, but by the time we get `PublishData` it is already gone).

However, I implemented extracting metadata from strings. It is not
guaranteed to work with all cases:

https://github.com/arturaz/mill/blob/e88240354d4ba417748aedfebda6f7c8b1c92d62/libs/javalib/src/mill/javalib/publish/PublishInfo.scala#L69-L101,
but it seems to work fine for the default publishing workflow, which is
what most people use.

For future reference, the whole publishing pipeline needs improvements,
as mentioned in https://github.com/com-lihaoyi/mill/issues/5538.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>