keytool functionalities for androidlib (#5473)
This PR replaces the use of the `keytool` CLI with a JVM-based
implementation that handles keystore creation, key pair and certificate
generation directly in code.

### Motivation

We want to be compatible with keystores created by Android Studio.
Gradle seems to use the `KeystoreHelper` class for this ([source
here](https://cs.android.com/android-studio/platform/tools/base/+/mirror-goog-studio-main:sdk-common/src/main/java/com/android/ide/common/signing/KeystoreHelper.java)),
which we looked into.

While we could’ve pulled that in as a dependency, it comes with some
Android/Gradle-specific stuff we don’t need. So for now, we’ve gone with
a custom implementation that does just what we need.

## Notes

We tried using BouncyCastle as the provider for the Keystore handling
but that resulted to unexpected behaviour with key passwords, so we just
use default provider instead.

### What this changes

- Removes the `keytool` call used for debug keystore creation
- Adds code (and Tests) to:
- Create a new keystore
- Add a key pair and certificate
- Save and load a created keystore
- Load a keystore created by AGP and verify the keypair

### What’s next

We'll add a command in a future PR to let users generate a release
keystore on their own.